requirements.txt
Pillow
requests
vercel-blob